FIX: Move proof interfaces in types package
