+++ data/slash/multipleHeaderProposalProof.proto
@@ -7,6 +7,7 @@ option (gogoproto.stable_marshaler_all) = true;
 
 import "slash.proto";
 import "github.com/gogo/protobuf/gogoproto/gogo.proto";
+import "github.com/ElrondNetwork/elrond-go-core/data/block/headerInfo.proto";
 
 message MultipleHeaderProposalProof {
   SlashingType   Type        = 1;
