+++ data/slash/slash.proto
@@ -6,7 +6,6 @@ option go_package = "slash";
 option (gogoproto.stable_marshaler_all) = true;
 
 import "github.com/gogo/protobuf/gogoproto/gogo.proto";
-import "github.com/ElrondNetwork/elrond-go-core/data/block/headerInfo.proto";
 
 enum SlashingType {
   None             = 0;
@@ -20,6 +19,3 @@ enum ThreatLevel {
   High   = 2;
 }
 
-message HeaderInfoList{
-  repeated HeaderInfo Headers = 1;
-}
